+++ server/lessons/theBeginnings/aNUakSSrvphpjtLn/scenes/main/q3/evts.js
@@ -1,0 +1,13 @@
+import * as a from "../../../../../../../src/lesson";
+return [
+    new LE_Bubble([
+        "Problem 3"
+    ],[
+        new BubbleTask("Task #1"),
+        new AddFileTask("main.js"),
+        new AddCode([
+            new CP_JS_Class("Particle")
+        ],null),
+        new BubbleTask("Problem 3 Finished!")
+    ])
+]